ml-service/ml-training-python/requirements-minimal.txt
# Minimal Requirements for ML Training Scheduler
# Only essential packages for training functionality

# Core ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Database Connections
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pymongo>=4.3.0
neo4j>=5.8.0
pymysql>=1.0.0
mysql-connector-python>=8.1.0

# Configuration Management
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0

# Logging
structlog>=23.0.0

# HTTP Requests
requests>=2.30.0

# Utilities
tqdm>=4.65.0
